requirements.txt
# ProQuants Professional Trading System Requirements
# Unified AI/ML/Neural Network System - Pure MT5 Integration

# Core Dependencies
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# AI/ML/Neural Network Stack
tensorflow>=2.10.0
keras>=2.10.0
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0

# Deep Learning Utilities
xgboost>=1.7.0
lightgbm>=4.0.0

# GUI Framework
tkinter
Pillow>=10.0.0
matplotlib>=3.7.0

# Data Processing & Analysis
pandas-ta>=0.3.14b0
QuantLib-Python>=1.31

# Utilities & Performance
requests>=2.31.0
pytz>=2023.3
python-dateutil>=2.8.2
numba>=0.57.0
psutil>=5.9.0

# Visualization
seaborn>=0.12.0
plotly>=5.15.0

# Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced Features
jupyter>=1.0.0
matplotlib>=3.7.0

# Build Dependencies
pyinstaller>=5.13.0
setuptools>=68.0.0
wheel>=0.41.0

# Note: ML-Enhanced Adaptive Levels System
# - Detects broker manipulation patterns
# - Optimizes fibonacci levels dynamically  
# - Adapts stop losses and entries based on market conditions
# - Counters algorithmic trading interference
